Modifica per non permettere di cancellare la prenotazione di altri utenti da parte di un'utente

+++ src/main/java/it/uniroma3/siw/my_restaurant/controller/ReservationController.java
@@ -23,29 +23,28 @@ import it.uniroma3.siw.my_restaurant.service.ReservationService;
 import it.uniroma3.siw.my_restaurant.service.UserService;
 import jakarta.validation.Valid;
 
-
 @Controller
 public class ReservationController {
 
-    @Autowired
-    private GlobalController globalController;
+	@Autowired
+	private GlobalController globalController;
 
-    @Autowired
-    private CredentialsService credentialsService;
+	@Autowired
+	private CredentialsService credentialsService;
 
-    @Autowired
-    private ReservationService reservationService;
+	@Autowired
+	private ReservationService reservationService;
 
-    @Autowired
-    private ReservationValidator reservationValidator;
+	@Autowired
+	private ReservationValidator reservationValidator;
 
 	@Autowired
 	private UserService userService;
 
-    @GetMapping("/reservation")
+	@GetMapping("/reservation")
 	public String getFormReservation(Model model) {
-        UserDetails userDetails = this.globalController.getUser();
-        Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
+		UserDetails userDetails = this.globalController.getUser();
+		Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
 
 		model.addAttribute("loggedUser", credentials);
 		model.addAttribute("reservation", new Reservation());
@@ -53,22 +52,23 @@ public class ReservationController {
 	}
 
 	@PostMapping("/reservation")
-	public String postFormReservation(@Valid @ModelAttribute("reservation") Reservation reservation, BindingResult bindingResult, Model model) {
+	public String postFormReservation(@Valid @ModelAttribute("reservation") Reservation reservation,
+			BindingResult bindingResult, Model model) {
 
-        UserDetails userDetails = this.globalController.getUser();
-        Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
+		UserDetails userDetails = this.globalController.getUser();
+		Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
 		model.addAttribute("loggedUser", credentials);
 
 		reservation.setUser(credentials.getUser());
-        if (reservation.getPlace().equals(Reservation.INTERNO_POSTO)) 
+		if (reservation.getPlace().equals(Reservation.INTERNO_POSTO))
 			reservation.setPlace(Reservation.INTERNO_POSTO);
-		else 
+		else
 			reservation.setPlace(Reservation.ESTERNO_POSTO);
 
-        this.reservationValidator.validate(reservation, bindingResult);
+		this.reservationValidator.validate(reservation, bindingResult);
 
 		if (!bindingResult.hasErrors()) {
-            this.reservationService.save(reservation);
+			this.reservationService.save(reservation);
 			return "/reservationSuccessful.html";
 		}
 
@@ -79,7 +79,7 @@ public class ReservationController {
 	public String getShowListAllReservationUser(Model model) {
 		UserDetails userDetails = this.globalController.getUser();
 		Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
-		
+
 		List<Reservation> allReservation = this.reservationService.findAllReservationsOfUser(credentials.getUser());
 		model.addAttribute("reservationList", allReservation);
 		model.addAttribute("loggedUser", credentials);
@@ -93,8 +93,8 @@ public class ReservationController {
 		Reservation reservation = this.reservationService.findById(reservationId);
 
 		if (!reservation.getUser().getId().equals(credentials.getUser().getId())) {
-        	return "error/403"; 
-    	}
+			return "error/403";
+		}
 
 		this.reservationService.delete(reservationId);
 		return "redirect:/user/allReservation";
@@ -108,55 +108,54 @@ public class ReservationController {
 		Reservation reservation = this.reservationService.findById(reservationId);
 
 		if (!reservation.getUser().getId().equals(credentials.getUser().getId())) {
-        	return "error/403"; 
-    	}
-		
+			return "error/403";
+		}
+
 		model.addAttribute("loggedUser", credentials);
 		model.addAttribute("reservation", reservation);
 
 		return "/updateReservation.html";
 	}
-	
+
 	@PostMapping("/user/{id}/updateReservation")
-	public String postUpdateReservationForm(@Valid @ModelAttribute("reservation") Reservation updateReservation, 
-				BindingResult bindingResult, @PathVariable("id") Long reservationId, Model model) {
+	public String postUpdateReservationForm(@Valid @ModelAttribute("reservation") Reservation updateReservation,
+			BindingResult bindingResult, @PathVariable("id") Long reservationId, Model model) {
 
 		UserDetails userDetails = this.globalController.getUser();
-        Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
+		Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
 		model.addAttribute("loggedUser", credentials);
 
-        if (updateReservation.getPlace().equals(Reservation.INTERNO_POSTO)) 
+		if (updateReservation.getPlace().equals(Reservation.INTERNO_POSTO))
 			updateReservation.setPlace(Reservation.INTERNO_POSTO);
-		else 
+		else
 			updateReservation.setPlace(Reservation.ESTERNO_POSTO);
 
-        this.reservationValidator.validate(updateReservation, bindingResult);
+		this.reservationValidator.validate(updateReservation, bindingResult);
 
 		if (!bindingResult.hasErrors()) {
 			updateReservation.setUser(credentials.getUser());
-            this.reservationService.update(updateReservation, reservationId);
+			this.reservationService.update(updateReservation, reservationId);
 			return "/reservationUpdateSuccessful.html";
 		}
 
 		return "/updateReservation.html";
 	}
 
-//	_______ PARTE ADMIN ________
+	// _______ PARTE ADMIN ________
 
 	@GetMapping("/admin/allReservation")
 	public String getShowListAllReservations(Model model) {
 		UserDetails userDetails = this.globalController.getUser();
 		Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
 
-		if(credentials.getRole().equals(ADMIN_ROLE))
-		{
+		if (credentials.getRole().equals(ADMIN_ROLE)) {
 			List<Reservation> allReservation = this.reservationService.findAllReservations();
 			model.addAttribute("reservationList", allReservation);
 			model.addAttribute("loggedUser", credentials);
 
 			return "/admin/showAllReservation.html";
 		}
-		
+
 		return "redirect:/login";
 	}
 
@@ -165,7 +164,6 @@ public class ReservationController {
 		this.reservationService.delete(reservationId);
 		return "redirect:/admin/allReservation";
 	}
-	
 
 	@GetMapping("/admin/AllUsers/{id}/reservationUser")
 	public String getAdminShowUserReservation(@PathVariable("id") Long userId, Model model) {
@@ -182,20 +180,27 @@ public class ReservationController {
 
 		return "/admin/showUserReservation.html";
 	}
-	
+
 	@PostMapping("/admin/AllUsers/{uid}/reservationUser/{rid}/delete")
 	public String postUserReservationDelete(@PathVariable("rid") Long reservationId, @PathVariable("uid") Long userId) {
-		this.reservationService.delete(reservationId);
-		return "redirect:/admin/AllUsers/" + userId + "/reservationUser";
+		UserDetails userDetails = this.globalController.getUser();
+		Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
+		if (this.reservationService.findById(reservationId).getUser().getId() == credentials.getId()) {
+			this.reservationService.delete(reservationId);
+			return "redirect:/admin/AllUsers/" + userId + "/reservationUser";
+		}
+
+		return "redirect:/";
+
 	}
-    
+
 	@GetMapping("/admin/{id}/updateReservation")
 	public String getUpdateReservationFormAdmin(@PathVariable("id") Long reservationId, Model model) {
 		UserDetails userDetails = this.globalController.getUser();
 		Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
 
 		Reservation reservation = this.reservationService.findById(reservationId);
-		
+
 		model.addAttribute("loggedUser", credentials);
 		model.addAttribute("reservation", reservation);
 
@@ -203,23 +208,23 @@ public class ReservationController {
 	}
 
 	@PostMapping("/admin/{id}/updateReservation")
-	public String postUpdateReservationFormAdmin(@Valid @ModelAttribute("reservation") Reservation updateReservation, 
-				BindingResult bindingResult, @PathVariable("id") Long reservationId, Model model) {
+	public String postUpdateReservationFormAdmin(@Valid @ModelAttribute("reservation") Reservation updateReservation,
+			BindingResult bindingResult, @PathVariable("id") Long reservationId, Model model) {
 
 		UserDetails userDetails = this.globalController.getUser();
-        Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
+		Credentials credentials = this.credentialsService.getCredentials(userDetails.getUsername());
 		model.addAttribute("loggedUser", credentials);
 
-        if (updateReservation.getPlace().equals(Reservation.INTERNO_POSTO)) 
+		if (updateReservation.getPlace().equals(Reservation.INTERNO_POSTO))
 			updateReservation.setPlace(Reservation.INTERNO_POSTO);
-		else 
+		else
 			updateReservation.setPlace(Reservation.ESTERNO_POSTO);
 
-        this.reservationValidator.validate(updateReservation, bindingResult);
+		this.reservationValidator.validate(updateReservation, bindingResult);
 
 		if (!bindingResult.hasErrors()) {
 			updateReservation.setUser(credentials.getUser());
-            this.reservationService.update(updateReservation, reservationId);
+			this.reservationService.update(updateReservation, reservationId);
 			return "/reservationUpdateSuccessful.html";
 		}
